a new pallet is born.
