+++ primitives/sidechain/src/lib.rs
@@ -1,0 +1,52 @@
+/*
+	Copyright 2021 Integritee AG and Supercomputing Systems AG
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+		http://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+
+*/
+
+//!Primitives for the sidechain
+#![cfg_attr(not(feature = "std"), no_std)]
+use codec::{Decode, Encode};
+use scale_info::TypeInfo;
+use sp_core::H256;
+use sp_io::hashing::blake2_256;
+use sp_std::prelude::*;
+
+#[cfg(feature = "std")]
+use serde::{Deserialize, Serialize};
+
+pub type ShardIdentifier = H256;
+
+#[derive(PartialEq, Eq, Clone, Encode, Decode, Debug, Copy, Default, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub struct SidechainHeader {
+	/// The parent hash.
+	pub parent_hash: H256,
+
+	/// The block number.
+	pub block_number: u64,
+
+	/// The Shard id.
+	pub shard_id: ShardIdentifier,
+
+	/// The payload hash.
+	pub block_data_hash: H256,
+}
+
+impl SidechainHeader {
+	/// get the `blake2_256` hash of the header.
+	pub fn hash(&self) -> H256 {
+		self.using_encoded(blake2_256).into()
+	}
+}
