+++ teerex/src/weights.rs
@@ -54,7 +54,6 @@ pub trait WeightInfo {
 	fn unregister_enclave() -> Weight;
 	fn call_worker() -> Weight;
 	fn confirm_processed_parentchain_block() -> Weight;
-	fn confirm_proposed_sidechain_block() -> Weight;
 }
 
 /// Weights for pallet_teerex using the Integritee parachain node and recommended hardware.
@@ -78,11 +77,6 @@ impl<T: frame_system::Config> WeightInfo for IntegriteeWeight<T> {
 			.saturating_add(T::DbWeight::get().reads(1 as Weight))
 			.saturating_add(T::DbWeight::get().writes(2 as Weight))
 	}
-	fn confirm_proposed_sidechain_block() -> Weight {
-		(46_200_000 as Weight)
-			.saturating_add(T::DbWeight::get().reads(1 as Weight))
-			.saturating_add(T::DbWeight::get().writes(2 as Weight))
-	}
 }
 
 // For tests
@@ -102,11 +96,6 @@ impl WeightInfo for () {
 	}
 	fn confirm_processed_parentchain_block() -> Weight {
 		(46_900_000 as Weight)
-			.saturating_add(RocksDbWeight::get().reads(1 as Weight))
-			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
-	}
-	fn confirm_proposed_sidechain_block() -> Weight {
-		(46_200_000 as Weight)
 			.saturating_add(RocksDbWeight::get().reads(1 as Weight))
 			.saturating_add(RocksDbWeight::get().writes(2 as Weight))
 	}
